Skip time zone in test by using Calendar, not Date
A `java.util.Date` has the local time zone embedded in it, so
instantiating it with the seconds since the epoch will always produce a
local time that varies.

Conversely, `java.util.Calendar` works differently; it always gives back
a date (using `Calendar.getTime()`) that matches the numbers passed in,
even if the time zone does not match. We can use this to ignore the time
zone component of the date but get a `Date` instance for the exact
desired date and time.